+++ packages/algo/CHANGELOG.md
@@ -1,3 +1,11 @@
+## [0.13.1] - 2025-10-05
+
+### Changed
+
+- Add `BitGrid.getSize`
+- Add `BitGrid.setAll`
+- Add `BitGrid.getByPosition`
+
 ## [0.13.0] - 2025-10-05
 
 ### Changed